+++ src/main/java/com/searchengine/solr/Controller/SolrCertificateController.java
@@ -1,7 +1,7 @@
 package com.searchengine.solr.Controller;
 
 import com.searchengine.solr.Model.Certification.Root;
-import com.searchengine.solr.Service.ServiceImpl.SolrCertificateService;
+import com.searchengine.solr.Service.ServiceImpl.SolrCertificateServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -15,10 +15,10 @@ import java.util.List;
 public class SolrCertificateController {
 
     @Autowired
-    SolrCertificateService solrCertificateService;
+    SolrCertificateServiceImpl solrCertificateServiceImpl;
 
     @RequestMapping("get-certificates-by-name")
     List<Root> getCertificatesByName(@RequestParam String name) throws IOException {
-        return  solrCertificateService.getCertificatesByName(name);
+        return  solrCertificateServiceImpl.getCertificatesByName(name);
     }
 }
