+++ src/main/java/com/searchengine/solr/Service/ServiceImpl/SolrCertificateServiceImpl.java
@@ -6,8 +6,9 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.searchengine.solr.Model.Certification.Root;
 import com.searchengine.solr.Model.CountObj;
 import com.searchengine.solr.Repository.SolrCertificateRepository;
-import com.searchengine.solr.Service.ICertificateService;
+import lombok.SneakyThrows;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.*;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
@@ -17,7 +18,7 @@ import java.util.*;
 import java.util.stream.Collectors;
 
 @Service
-public class SolrCertificateService implements ICertificateService {
+public class SolrCertificateServiceImpl implements com.searchengine.solr.Service.SolrCertificateService {
 
     @Autowired
     SolrCertificateRepository solrCertificateRepository;
@@ -25,16 +26,43 @@ public class SolrCertificateService implements ICertificateService {
     @Autowired
     RestTemplate restTemplate;
 
+    @Value("${solr.port}")
+    private String solrPort;
+
     @Override
     public List<Root> getCertificatesByName(String name) {
-
-        List<Root> rootList = solrCertificateRepository.findByName(name);
-        return rootList;
+        try {
+            List<Root> rootList = solrCertificateRepository.findByName(name);
+            return rootList;
+        }
+        catch (Exception e){
+            throw e;
+        }
     }
 
     @Override
     public String getCertifiedHashersCount(String name) throws IOException {
-        return null;
+        ObjectMapper objectMapper = new ObjectMapper();
+        List<Root> roots = solrCertificateRepository.findByName(name);
+        String rootId = "";
+        for(Root root : roots)
+            rootId = root.get_root_();
+        System.out.println("rootId ::::" + rootId);
+        HttpHeaders headers = new HttpHeaders();
+        headers.setAccept(Arrays.asList(MediaType.APPLICATION_JSON));
+        HttpEntity<String> entity = new HttpEntity<String>(headers);
+        String url = solrPort+"/solr/certifications/select?q=id:"+rootId+"*&fl=count&rows=500";
+        ResponseEntity<String> response = restTemplate.exchange(url,
+                HttpMethod.GET, new HttpEntity<String>(headers), String.class);
+        System.out.println(response.getBody());
+        LinkedHashMap<String,String> l = new LinkedHashMap<>();
+
+        JsonNode jsonNode = objectMapper.readTree(response.getBody()).get("response").get("docs");
+
+        ArrayList<HashMap<String,List<Integer>>> arrayList = objectMapper.convertValue(jsonNode, ArrayList.class);
+        Integer count = arrayList.stream().filter(e -> !e.isEmpty()).filter(e -> e.containsKey("count")).collect(Collectors.toList()).get(0).get("count").get(0);
+        String hashersCount = "Number of certified hashers : "+ count.toString();
+        return hashersCount;
     }
 
     @Override
@@ -45,7 +73,7 @@ public class SolrCertificateService implements ICertificateService {
         HttpEntity<String> entity = new HttpEntity<String>(headers);
         HashMap<String, String> countMap = new HashMap<>();
         ArrayList<CountObj> finalResponse = new ArrayList<>();
-        String url = "http://localhost:8984/solr/certifications/select?fl=count,id&indent=true&q.op=OR&q=id:*/certifiedHashersSummary" + "?" + "&rows=1000&useParams=";
+        String url = solrPort+"/solr/certifications/select?fl=count,id&indent=true&q.op=OR&q=id:*/certifiedHashersSummary" + "?" + "&rows=1000&useParams=";
         ResponseEntity<String> response = restTemplate.exchange(url,
                 HttpMethod.GET, new HttpEntity<String>(headers), String.class);
         LinkedHashMap<String, String> l = new LinkedHashMap<>();
@@ -57,7 +85,7 @@ public class SolrCertificateService implements ICertificateService {
         ArrayList<HashMap<HashMap<String, String>, HashMap<String, String>>> arrayList = objectMapper.convertValue(jsonNode, ArrayList.class);
         for (HashMap c : arrayList) {
             String id = Arrays.stream(c.get("id").toString().split("/")).collect(Collectors.toList()).get(0);
-            String uri = "http://localhost:8984/solr/certifications/select?fl=name&indent=true&q.op=OR&q=id:" + id + "&useParams=";
+            String uri = solrPort+"/solr/certifications/select?fl=name&indent=true&q.op=OR&q=id:" + id + "&useParams=";
             ResponseEntity<String> nameResponse = restTemplate.exchange(uri,
                     HttpMethod.GET, new HttpEntity<String>(headers), String.class);
             JsonNode jsonNode1 = null;
