+++ src/main/java/com/searchengine/solr/config/SolrConfig.java
@@ -2,6 +2,7 @@ package com.searchengine.solr.config;
 
 import org.apache.solr.client.solrj.SolrClient;
 import org.apache.solr.client.solrj.impl.CloudSolrClient;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.solr.core.SolrTemplate;
@@ -12,11 +13,13 @@ import java.util.Optional;
 
 @Configuration
 public class SolrConfig {
+    @Value("${zookeeper.port}")
+    private String zookeeperPort;
 
     @Bean
     public CloudSolrClient solrClient() {
         final List<String> zkServers = new ArrayList<>();
-        zkServers.add("localhost:9984");
+        zkServers.add(zookeeperPort);
         return new CloudSolrClient.Builder(zkServers, Optional.empty()).build();
     }
